Fix newline bugs
The line variables end in a newline, which are now removed with
replace(). float() will remove these newlines, but this is to prevent
errors in case some other function other than float() is used in the
future.

replace() is preferred over cutting out the last character because we
cannot assume that the last line of every file ends in a newline.